server/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
contextvars==2.4
Django==4.1.4
flake8==6.0.0
frozenlist==1.3.3
gunicorn==20.1.0
ibm-cloud-sdk-core==3.16.0
ibm-watson==6.1.0
ibmcloudant==0.3.1
idna==3.4
immutables==0.19
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.3
Pillow==9.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
websocket-client==1.1.0
yarl==1.8.2
